vantbot-scripts/requirements.txt
ollama
flask
slack_sdk
requests
psycopg2-binary
pdfplumber
pillow
pandas
openpyxl
markdown
html-slacker